added inventory & boutique (gateway+routes)

+++ boutique/src/main/java/fr/pantheonsorbonne/ufr27/miage/dao/ProductDAO.java
@@ -23,7 +23,7 @@ public class ProductDAO {
                 .getResultList();
     }
 
-    public Product findById(Long id) {
+    public Product findById(Integer id) {
         return em.find(Product.class, id);
     }
 
@@ -35,7 +35,7 @@ public class ProductDAO {
         em.merge(product);
     }
 
-    public void delete(Long id) {
+    public void delete(Integer id) {
         Product product = findById(id);
         if (product != null) {
             em.remove(product);
